feat: set up squid sdk, split out deposit form
