+++ apps/trading/.env
@@ -44,5 +44,5 @@ NX_CHARTING_LIBRARY_HASH=PDjWaqPFndDp+LCvqbKvntWriaqNzNpZ5i9R/BULzCg=
 
 NX_VEGA_NETWORKS={\"MAINNET\":\"https://console.vega.xyz\",\"TESTNET\":\"https://console.fairground.wtf\",\"STAGNET1\":\"https://trading.stagnet1.vega.rocks\"}
 NX_SQUID_INTEGRATOR_ID=vega-swap-widget
-NX_SQUID_API_URL=https://api.squidrouter.com
+NX_SQUID_API_URL=https://v2.api.squidrouter.com
